irontrack/pyproject.toml
[project]
name = "irontrack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "pydantic==2.9.2",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.12",
    "sqlalchemy==2.0.36",
    "uvicorn[standard]==0.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["irontrack"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
